Rename package from s3 to rgw
