+++ internal/rgw/client_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package s3
+package rgw
 
 import (
 	"testing"
